+++ examples/src/vanillaBundle.js
@@ -1,0 +1,6 @@
+import Dice from '../../vanillaSource';
+
+const dice = new Dice();
+const htmlNode = dice.render();
+
+document.getElementById('vanilla-game').appendChild(htmlNode);