+++ index.js
@@ -1,3 +1,0 @@
-import GenericDice from './components/generic-dice/generic-dice';
-
-export default GenericDice;