+++ reactSource.js
@@ -1,0 +1,3 @@
+import GenericDice from './src/react/generic-dice/generic-dice';
+
+export default GenericDice;