+++ vanillaSource.js
@@ -1,0 +1,3 @@
+import GenericDiceVanilla from './src/vanilla/generic-dice-vanilla';
+
+export default GenericDiceVanilla;