+++ CHANGELOG.md
@@ -3,8 +3,17 @@ A tag is created for each release from the master branch.
 
 ### Current version being developped
 
+- 1.2.0:
+   - Review dependencies with npm-check (issue 12)
+   - Programmatically trigger a new roll from Javascript (issue 10)
+   - After .roll() is called there is a delay before the component is updated back to its non rolling state (issue 16)
+   - Make it possible to change the duration of the rolling (issue 23)
 
 ### Release history
+- 1.1.5:
+   - Support bulding plain Javascript (issue 21)
+   - Refactor dice-face and dice-shadow into the main generic-dice component (issue 29)
+   - Fix installing generic-dice as an npm module (issue 30)
 - 1.1.4 Unit testing.
 - 1.1.3 Fixed the shadow animation.
 - 1.1.2 Refactor to setup a working npm module.